Registro de Movimientos Inicial

+++ src/main/java/com/myfinance/entities/Cuenta.java
@@ -1,5 +1,6 @@
 package com.myfinance.entities;
 
+import javafx.beans.property.SimpleDoubleProperty;
 import javafx.beans.property.SimpleIntegerProperty;
 import javafx.beans.property.SimpleStringProperty;
 
@@ -8,16 +9,16 @@ public class Cuenta {
     private int cuentaID;
     private String nombreUsuario;
     private String nombreCuenta;
-    private int saldo;
+    private double saldo;
 
-    public Cuenta(int cuentaID, String nombreUsuario, String nombreCuenta, int saldo) {
+    public Cuenta(int cuentaID, String nombreUsuario, String nombreCuenta, double saldo) {
         this.cuentaID = cuentaID;
         this.nombreUsuario = nombreUsuario;
         this.nombreCuenta = nombreCuenta;
         this.saldo = saldo;
     }
 
-    public Cuenta(String nombreUsuario, String nombreCuenta, int saldo) {
+    public Cuenta(String nombreUsuario, String nombreCuenta, double saldo) {
         this.nombreUsuario = nombreUsuario;
         this.nombreCuenta = nombreCuenta;
         this.saldo = saldo;
@@ -47,11 +48,11 @@ public class Cuenta {
         this.nombreCuenta = nombreCuenta;
     }
 
-    public int getSaldo() {
+    public double getSaldo() {
         return saldo;
     }
 
-    public void setSaldo(int saldo) {
+    public void setSaldo(double saldo) {
         this.saldo = saldo;
     }
 
@@ -59,7 +60,7 @@ public class Cuenta {
         return new SimpleStringProperty(this.nombreCuenta);
     }
 
-    public SimpleIntegerProperty getPropertySaldo(){
-        return new SimpleIntegerProperty(this.saldo);
+    public SimpleDoubleProperty getPropertySaldo(){
+        return new SimpleDoubleProperty(this.saldo);
     }
 }
